+++ callflow/__init__.py
@@ -1,1 +1,3 @@
 ## callflow.__init__.py
+
+from .datastructures.graphframe import GraphFrame
