+++ callflow/modules/auxiliary.py
@@ -14,7 +14,7 @@ import json
 import networkx as nx
 from ast import literal_eval as make_tuple
 import numpy as np
-from .gradients import Gradients
+#from .gradients import Gradients
 from .boxplot import BoxPlot
 from callflow.utils.logger import Log
 from callflow.utils.timer import Timer
