+++ callflow/utils/__init__.py
@@ -1,5 +1,5 @@
 from .df import *
-from .graph import *
+#from .graph import *
 from .hatchet import *
 from .logger import *
 from .networkx import *
